+++ packages/gamut/src/Form/FormGroupDescription.tsx
@@ -1,6 +1,6 @@
 import { system } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 
-import { formBaseStyles } from './styles/shared-system-props';
+import { formBaseStyles } from './styles';
 
 export const FormGroupDescription = styled.div(system.css(formBaseStyles));
