+++ packages/gamut/src/Form/SelectDropdown.tsx
@@ -1,6 +1,5 @@
 import { ArrowChevronDownIcon } from '@codecademy/gamut-icons';
-import { theme } from '@codecademy/gamut-styles';
-import { css } from '@emotion/react';
+import { useTheme } from '@emotion/react';
 import React, {
   ReactNode,
   SelectHTMLAttributes,
@@ -18,12 +17,14 @@ import ReactSelect, {
 
 import { SelectComponentProps } from './Select';
 import {
-  colorStates,
-  conditionalBorderStyles,
-  formDropdownStyles,
-  formFieldStyles,
-} from './styles/shared';
-import { conditionalStyleProps } from './styles/shared-system-props';
+  conditionalBorderStates,
+  dropdownBorderStates,
+  dropdownBorderStyles,
+  optionBackground,
+  placeholderColor,
+  selectDropdownStyles,
+  textColor,
+} from './styles';
 import { parseOptions } from './utils';
 
 const { DropdownIndicator, SelectContainer } = SelectDropdownElements;
@@ -53,7 +54,10 @@ type CustomContainerProps = ContainerProps<OptionStrict, false> & {
 const ChevronDropdown = (props: IndicatorProps<OptionTypeBase, false>) => {
   return (
     <DropdownIndicator {...props}>
-      <ArrowChevronDownIcon size={16} />
+      <ArrowChevronDownIcon
+        size={16}
+        color={props.isDisabled ? 'text-disabled' : 'text'}
+      />
     </DropdownIndicator>
   );
 };
@@ -70,85 +74,9 @@ const CustomContainer = ({ children, ...rest }: CustomContainerProps) => {
   );
 };
 
-const selectBaseStyles = ({
-  error,
-  activated,
-  isFocused,
-  isDisabled,
-}: conditionalStyleProps) => css`
-  ${formFieldStyles}
-  ${conditionalBorderStyles({ error, activated, isFocused, isDisabled })}
-  line-height: ${theme.lineHeight.base};
-  display: flex;
-`;
-
-const errorColorState = (error: boolean) => {
-  const color = error ? colorStates.error.color : colorStates.base.color;
-  return color;
-};
-
-const customStyles: StylesConfig<OptionTypeBase, false> = {
-  container: (provided, state) => ({
-    ...provided,
-    pointerEvents: 'visible',
-    cursor: state.selectProps.isSearchable ? 'text' : 'pointer',
-    width: '100%',
-    minWidth: '7rem',
-  }),
-
-  control: (provided, state) => ({
-    ...selectBaseStyles({
-      error: state.selectProps.error,
-      activated: state.selectProps.activated,
-      isFocused: state.isFocused,
-      isDisabled: state.isDisabled,
-    }),
-  }),
-
-  dropdownIndicator: (provided, state) => ({
-    color: errorColorState(state.selectProps.error),
-    display: 'flex',
-    padding: '0',
-    pointerEvents: 'none',
-  }),
-
-  input: (provided, state) => ({
-    padding: '0',
-    margin: '0',
-  }),
-
-  menu: (provided, state) => ({
-    ...provided,
-    ...formDropdownStyles(state.selectProps.error),
-  }),
-
-  option: (provided, state) => ({
-    padding: '14px 11px 14px 11px',
-    cursor: 'pointer',
-    backgroundColor:
-      (state.isSelected && colorStates.dropdown.selected.backgroundColor) ||
-      (state.isFocused && colorStates.dropdown.focused.backgroundColor) ||
-      'transparent',
-    '&:hover': {
-      backgroundColor: colorStates.dropdown.focused.backgroundColor,
-    },
-  }),
-
-  singleValue: (provided, state) => ({
-    color: errorColorState(state.selectProps.error),
-    display: 'flex',
-  }),
-
-  valueContainer: (provided) => ({
-    ...provided,
-    padding: 0,
-  }),
-};
-
 const defaultProps = {
   name: undefined,
   isMulti: false,
-  styles: customStyles,
   components: {
     DropdownIndicator: ChevronDropdown,
     IndicatorSeparator: () => null,
@@ -180,6 +108,73 @@ export const SelectDropdown: React.FC<SelectDropdownProps> = ({
     setActivated(true);
   };
 
+  const theme = useTheme();
+
+  const memoizedStyles: StylesConfig<OptionTypeBase, false> = useMemo(() => {
+    return {
+      container: (provided, state) => ({
+        ...provided,
+        pointerEvents: 'visible',
+        cursor: state.selectProps.isSearchable ? 'text' : 'pointer',
+        width: '100%',
+        minWidth: '7rem',
+      }),
+
+      control: (provided, state) => {
+        return {
+          ...selectDropdownStyles({ theme }),
+          ...conditionalBorderStates({
+            isFocused: state.isFocused,
+            isDisabled: state.isDisabled,
+            error: state.selectProps.error,
+            activated: state.selectProps.activated,
+            theme,
+          }),
+        };
+      },
+
+      dropdownIndicator: (provided, state) => ({
+        color: 'currentColor',
+        display: 'flex',
+        padding: '0',
+        pointerEvents: 'none',
+      }),
+
+      input: (provided, state) => ({
+        ...textColor({ theme }),
+        padding: '0',
+        margin: '0',
+      }),
+
+      menu: (provided, state) => ({
+        ...provided,
+        ...dropdownBorderStyles({ theme }),
+        ...dropdownBorderStates({ error: state.selectProps.error, theme }),
+      }),
+
+      placeholder: (provided, state) => ({
+        ...provided,
+        ...placeholderColor({ theme }),
+      }),
+
+      option: (provided, state) => ({
+        padding: '14px 11px 14px 11px',
+        cursor: 'pointer',
+        ...optionBackground(state.isSelected, state.isFocused)({ theme }),
+      }),
+
+      singleValue: (provided, state) => ({
+        ...textColor({ theme }),
+        display: 'flex',
+      }),
+
+      valueContainer: (provided) => ({
+        ...provided,
+        padding: 0,
+      }),
+    };
+  }, [theme]);
+
   const selectOptions = useMemo(() => {
     return parseOptions({ options, id });
   }, [options, id]);
@@ -204,6 +199,7 @@ export const SelectDropdown: React.FC<SelectDropdownProps> = ({
       isDisabled={disabled}
       options={selectOptions}
       placeholder={placeholder}
+      styles={memoizedStyles}
       isSearchable={isSearchable}
       {...rest}
     />
