--- packages/gamut/src/Form/styles/FormGroupDescription.module.scss
@@ -1,7 +1,0 @@
-@import "variables";
-
-.FormGroupDescription {
-  margin-bottom: 0.5rem;
-  font-weight: normal;
-  line-height: 1.25rem;
-}
