--- packages/gamut/src/Form/styles/_variables.scss
@@ -1,7 +1,0 @@
-@import "~@codecademy/gamut-styles/utils";
-$color-border: $color-gray-200;
-$item-background: $color-white;
-$item-background-hover: $color-gray-200;
-$border-radius: 2px;
-$transition-time: 0.15s;
-$form-padding: 0.8rem;
