+++ packages/styleguide/stories/Atoms/FormInputs/SelectDropdown.stories.mdx
@@ -4,6 +4,8 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
 
 import { PropsTable } from '~styleguide/blocks';
 
+import { SelectDropdownExample } from './examples';
+
 export const fruitOptions = [
   'Apple',
   'Banana',
@@ -218,3 +220,9 @@ Some of our forms require special props be passed directly onto the `<input/>` c
     )}
   </Story>
 </Canvas>
+
+### ColorMode
+
+Our SelectDropdown is responsive to ColorMode!
+
+<SelectDropdownExample />
